split element from particle
